Add a TreeGroup type
To help split up dialog into different files of related nodes.

Note that allNodes isn't traversed at this time, so the separate
type may not be necessary. All we're doing is providing a place
to define them so they are instantiated and then the rest is done
through node links.